Took care that the toolset compiles again by renaming a size_t into an unsigned int.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8425 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/source/memory.c
@@ -1998,7 +1998,7 @@ ATerm AT_removeAnnotations(ATerm t)
  * Change one argument of an application.
  */
 
-ATermAppl ATsetArgument(ATermAppl appl, ATerm arg, size_t n)
+ATermAppl ATsetArgument(ATermAppl appl, ATerm arg, unsigned int n)
 {
   unsigned int i, arity;
   Symbol sym = ATgetSymbol(appl);
